Further abstract activity types
init() is a bit broken that even ControlCenterV2 doesn't use it properly. This kinda breaks what we want to do for Fresh Wearable too so make this better.

Further abstract activity types so we can instead inherit two common activity types (with or without action bar) with the same logic on language or theme change. We have less code on the theme change since we have upped the dark theme logic to what the UI library provides.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>