Make app notification forwarding settings simpler and per-device
This is the first step into making options on forwarding notifications from the phone to a connected device simpler to make the UX better for Fresh Wearable. This also makes the notification settings enabled on a per-device basis, allowing users to enable notification forwarding for specific apps on a per-device basis.

Signed-off-by: John Vincent Corcega <git@tenseventyseven.xyz>